GROW-979 dynamicQueryCount returns long while we need an int

+++ modules/social-activity-customizer/social-activity-customizer-service/src/main/java/com/liferay/social/activity/customizer/service/persistence/impl/CustomSocialActivitySetFinderImpl.java
@@ -63,10 +63,10 @@ public class CustomSocialActivitySetFinderImpl
 					RestrictionsFactoryUtil.eq("userId", userId),
 					RestrictionsFactoryUtil.eq("classNameId", classNameId)));
 
-			int count = SocialActivitySetLocalServiceUtil.dynamicQueryCount(
+			long count = SocialActivitySetLocalServiceUtil.dynamicQueryCount(
 				activitySetQuery);
 
-			return count;
+			return Math.toIntExact(count);
 		}
 		catch (Exception e) {
 			try {
@@ -159,10 +159,10 @@ public class CustomSocialActivitySetFinderImpl
 					RestrictionsFactoryUtil.in(
 						"type", ArrayUtil.toArray(types))));
 
-			int count = SocialActivitySetLocalServiceUtil.dynamicQueryCount(
+			long count = SocialActivitySetLocalServiceUtil.dynamicQueryCount(
 				activitySetQuery);
 
-			return count;
+			return Math.toIntExact(count);
 		}
 		catch (Exception e) {
 			try {
